EQ2-157 relocate all tests in RootControllerTest that belong to the GestionnaireRootController to GestionnaireRootControllerTest
